LicensePlate_RESTAPI/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
Django==4.0.4
djangorestframework==3.13.1
numpy==1.22.4
opencv-python==4.5.5
Pillow==9.1.1
pip==21.2.4
pytz==2022.1
setuptools==61.2.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
wheel==0.37.1
wincertstore==0.2
